api/requirements/base.txt

python-dotenv 
djangorestframework 
django-autoslug 
django-countries 
django-phonenumber-field 
drf-yasg 
Pillow 
djoser
argon2-cffi 
django-storages 
django-taggit 
django-filter 
django-celery-email 
cloudinary 
django-celery-beat 
redis 
celery 
flower